+++ MaxBot/Objects/Recipient.cs
@@ -5,11 +5,11 @@ namespace MaxBot.Objects;
 public class Recipient
 {
     [JsonPropertyName("chat_id")]
-    public int? ChatId { get; set; }
+    public long? ChatId { get; set; }
 
     [JsonPropertyName("chat_type")]
     public string ChatType { get; set; } = null!;
 
     [JsonPropertyName("user_id")]
-    public int? UserId { get; set; }
+    public long? UserId { get; set; }
 }
